Implement table_slice to Record Batch conversion
